+++ src/extension.ts
@@ -1,29 +1,44 @@
-import { commands, ExtensionContext, workspace } from 'vscode';
-import { ProjectDetails } from './common/ProjectDetails';
+import { commands, ExtensionContext, workspace, window } from 'vscode';
+import * as _ from "lodash";
 import { RepositoryView } from './views/repositoryView';
 import { TravisStatusBar } from './views/travisStatusBarItem';
+import { ActiveRepositorySingleton } from './common/ActiveRepositorySingleton';
 
 export function activate(context: ExtensionContext) {
 
-  const projectDetailsInstance = new ProjectDetails();
+  // initialise singleton
+  let singletonInstance: any;
+
+  if (workspace.rootPath) {
+    const singleton = ActiveRepositorySingleton.createInstance(context, workspace.rootPath);
+    singletonInstance = ActiveRepositorySingleton.getInstance();
+  }
+
   const repoView = new RepositoryView(context);
   const tree = repoView.initialise();
   const travisStatusBarInstance = new TravisStatusBar(context);
 
   const disposable = commands.registerCommand('extension.theTravisClient', () => tree);
-
+  // Configuration change trigger events
   const onChangeConfigurationDisposable = workspace.onDidChangeConfiguration((event) => {
     const onChangeBranches = event.affectsConfiguration('travisClient.branches');
     const onChangePro = event.affectsConfiguration('travisClient.pro');
     const onChangeOwner = event.affectsConfiguration('travisClient.owner');
-    if (onChangeBranches || onChangePro || onChangeOwner) {
+    const onChangeInterval = event.affectsConfiguration('travisClient.interval');
+    if (onChangeBranches || onChangePro || onChangeOwner || onChangeInterval) {
       commands.executeCommand('theTravisClient.refresh');
     }
   });
 
-  const setProToken = commands.registerCommand('theTravisClient.setProToken', () => projectDetailsInstance.setAuthToken(context, 'enterprise'));
+  const setProToken = commands.registerCommand('theTravisClient.setProToken', () => singletonInstance.setAuthToken('enterprise'));
 
-  const setToken = commands.registerCommand('theTravisClient.setToken', () => projectDetailsInstance.setAuthToken(context, 'community'));
+  const setToken = commands.registerCommand('theTravisClient.setToken', () => singletonInstance.setAuthToken('community'));
+
+  const autoRefresh = function () {
+    commands.executeCommand('theTravisClient.refresh');
+  };
+
+  const autoRefreshInterval = setInterval(autoRefresh, singletonInstance.interval());
 
   const refresh = commands.registerCommand('theTravisClient.refresh', () => {
     const repoView = new RepositoryView(context);
@@ -36,6 +51,7 @@ export function activate(context: ExtensionContext) {
   context.subscriptions.push(setProToken);
   context.subscriptions.push(setToken);
   context.subscriptions.push(refresh);
+
 }
 
 // this method is called when your extension is deactivated