+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 ## [Unreleased]
 - Add Pull requests build status
 
+## [0.0.8] - 2019-05-31
+### Added
+- Auto refresh after given interval _(default: 5 minutes)_
+  ```json
+  "travisClient.interval": 5
+  ```
+  
 ## [0.0.7] - 2019-05-29
 - Fixed extension activation problem